refactor(core): extract FFI settings parsers
